+++ src/experiences/base-experience/frame/experience-frame.ts
@@ -26,7 +26,7 @@ import {
 } from '@common/events/events';
 import {Iframe} from '@common/iframe/iframe';
 
-export const SDK_VERSION = '2.5.0';
+export const SDK_VERSION = '2.6.0';
 
 export abstract class BaseExperienceFrame<
     ExperienceContentOptions extends ContentOptions,