+++ src/experiences/dashboard-experience/dashboard-experience.ts
@@ -258,6 +258,14 @@ export class DashboardExperience extends BaseExperience<
         );
     };
 
+    setPreloadThemes = async (preloadThemes: string[]): Promise<ResponseMessage> => {
+        return this.send(
+            new EmbeddingMessageEvent(MessageEventName.PRELOAD_THEMES, {
+                PreloadThemes: preloadThemes,
+            })
+        );
+    };
+
     protected extractExperienceFromUrl = (url: string): IDashboardExperience => {
         const matches: Array<string> = /^https:\/\/[^/]+\/embed\/[^/]+\/dashboards\/([\w-]+)(\?|$)/i.exec(url) || [];
 
@@ -288,6 +296,12 @@ export class DashboardExperience extends BaseExperience<
         // if the resizeHeightOnSizeChangedEvent is true, upon receiving SIZE_CHANGED message, update the height of the iframe
         if (messageEvent.eventName === 'SIZE_CHANGED' && this.frameOptions.resizeHeightOnSizeChangedEvent) {
             metadata?.frame?.setAttribute?.('height', `${messageEvent?.message?.height}px`);
+        }
+        if (messageEvent.eventName === 'EXPERIENCE_INITIALIZED' && this.contentOptions?.themeOptions?.themeOverride) {
+            this.setThemeOverride(this.contentOptions.themeOptions.themeOverride);
+        }
+        if (messageEvent.eventName === 'EXPERIENCE_INITIALIZED' && this.contentOptions?.themeOptions?.preloadThemes) {
+            this.setPreloadThemes(this.contentOptions.themeOptions.preloadThemes);
         }
     };